Déplace pole_emploi_categorie_demandeur_emploi dans salarie.py
